Question Model renamed to Problem for consistency with rest of code.
Submitting a form no longer saves the solution text as a Problem object which incorrectly
gets rendered on the same page. New Solution Model has been added, and now submitting a solution
saves it as a Solution object in the database.

TO-DO: Come up with a Schema for matching Solutions with their respective Problem and students.
Currently no way touniquely identify solutions, which will be necessary for data processing
and analysis.